refactor: Move visualizations to separate files.
